# inesdata-map-editor-frontend/src/openapi/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "Service responsible for saving a data source mapping and generating the corresponding R2RML / RML mapping file",
    "title": "com.inesdata-map:mapper-backend",
    "version": "1.1.0-SNAPSHOT"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/data-sources": {
      "get": {
        "operationId": "listDataSources",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDataSourceDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List all data sources",
        "tags": [
          "DataSource"
        ]
      }
    },
    "/data-sources/{id}": {
      "delete": {
        "operationId": "deleteDataSource",
        "parameters": [
          {
            "description": "Data source identifier to delete",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Delete data source",
        "tags": [
          "DataSource"
        ]
      }
    },
    "/database-sources": {
      "post": {
        "operationId": "createDataBaseSource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataBaseSourceDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceDTO"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Create given data base source",
        "tags": [
          "DataBaseSource"
        ]
      }
    },
    "/database-sources/{id}": {
      "get": {
        "operationId": "getDataBaseSource",
        "parameters": [
          {
            "description": "Data source identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataBaseSourceDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Gets given data base source",
        "tags": [
          "DataBaseSource"
        ]
      },
      "put": {
        "operationId": "updateDataBaseSource",
        "parameters": [
          {
            "description": "Data base source identifier to update",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataBaseSourceDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Update given data base source",
        "tags": [
          "DataBaseSource"
        ]
      }
    },
    "/database-sources/{id}/table-column-names": {
      "get": {
        "description": "Retrieves a list of column names for the specified table name and a data source ID.",
        "operationId": "getColumnNames",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "name": "table",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get column names for a database table",
        "tags": [
          "DataBaseSource"
        ]
      }
    },
    "/database-sources/{id}/table-names": {
      "get": {
        "description": "Retrieves a list of table names for the specified data source ID.",
        "operationId": "getTableNames",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get table names for a database source",
        "tags": [
          "DataBaseSource"
        ]
      }
    },
    "/executions/{id}/files": {
      "get": {
        "operationId": "downloadFile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Download the specified file from the given execution",
        "tags": [
          "Execution"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "DataBaseSourceDTO": {
        "required": [
          "connectionString",
          "databaseType",
          "name",
          "password",
          "userName"
        ],
        "type": "object",
        "properties": {
          "connectionString": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "databaseType": {
            "type": "string",
            "enum": [
              "POSTGRESQL",
              "MYSQL",
              "H2",
              "MONGO"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "DATABASE",
              "FILE"
            ]
          },
          "userName": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "description": "The data base source to save"
      },
      "DataSourceDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "DATABASE",
              "FILE"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PageDataSourceDTO": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataSourceDTO"
            }
          },
          "empty": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
...
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}